requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==5.3.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
cryptography==43.0.1
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.120.1
filelock==3.20.0
flasgger==0.9.7.1
Flask==3.1.2
flask-sock==0.7.0
flatbuffers==25.9.23
fsspec==2024.6.1
google-auth==2.42.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.0.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.1.0
kubernetes==34.1.0
langchain==0.3.6
langchain-openai==0.2.9
logistro==2.0.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==2.2.0
MarkupSafe==3.0.3
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
modelscope==1.11.0
mpmath==1.3.0
mysql==0.0.3
mysqlclient==2.2.7
narwhals==2.10.0
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.0
onnxruntime==1.23.2
openai==2.14.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==24.1
pandas==2.3.3
pandas-stubs==2.3.2.250926
pillow==12.0.0
plotly==6.3.1
pluggy==1.6.0
posthog==5.4.0
protobuf==5.26.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyMySQL==1.1.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.28.0
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.49.1
sympy==1.14.0
tabulate==0.9.0
tenacity==8.2.3
tokenizers==0.22.1
tomli==2.3.0
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
types-pytz==2025.2.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
vanna==0.7.9
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
zipp==3.23.0
akshare==1.10.50
